Data Collection, Cleaning and Duration Curve
This code collects monthly electricity pricing data for 2015 available at 5-minutes interval within location information. The data for all months is then consolidated to develop the annual pricing data. Subsets of this data are created based on the NY ISO zones. 

Duration Curves are generated for each of the zones. From the duration curves, the following information in extracted:
1. Hours when the price of electricity is $0
2. Hours when the price of electricity is $1
3. Hours when the price of electricity is $2
4. Hours when the price of electricity is $5